Create ConfigManager class
