+++ desktopsort.py
@@ -3,7 +3,7 @@ from os import walk
 import shutil
 from classes.FileInfo import FileInfo
 from classes.SortedFolder import SortedFolder
-import json
+from classes.ConfigManager import ConfigManager
 
 
 def get_sorting_input_files(path: str) -> list[FileInfo]:
@@ -94,34 +94,14 @@ def sort_files(path: str, files: list[FileInfo], folders: list[SortedFolder]) ->
             print("-------------------------\n")
 
 
-def load_config_from_file(path: str) -> str:
-    with open(path, "r") as file:
-        return file.read()
-
-
-def deserialize_config_json(json_data_dict: dict) -> list[SortedFolder]:
-    sorted_folders: list[SortedFolder] = []
-    for folder in json_data_dict:
-        dir: SortedFolder = SortedFolder("", [])
-
-        dir.name = folder["name"]
-
-        associated_types = folder["associated_types"]
-        for type in associated_types:
-            dir.associated_types.append(type)
-
-        sorted_folders.append(dir)
-    return sorted_folders
-
-
 def main() -> None:
     # path = desktop = os.path.join(os.path.join(os.path.expanduser('~')), 'Desktop') + '\\'
     path: str = "..\\ExampleDesktop"
 
     output_folders: SortedFolder = []
-    json_data_string = load_config_from_file("./config.json")
-    json_data = json.loads(json_data_string)
-    output_folders = deserialize_config_json(json_data)
+    config_manager: ConfigManager = ConfigManager
+    json_data_string = config_manager.load_config_from_file("./config.json")
+    output_folders = config_manager.deserialize_config_json(json_data_string)
     create_output_folders(path, output_folders)
 
     input_files: FileInfo = []
@@ -129,7 +109,6 @@ def main() -> None:
     input_files = exclude_sorting_files(input_files, output_folders)
 
     sort_files(path, input_files, output_folders)
-
 
 if __name__ == "__main__":
     main()
